filterx: introduce FilterXJIT
FilterXJIT is a Facade pattern, encapsulating everything needed for
FilterX to be a (JIT) compiled rather than an interpreted language:

- methods to generate LLVM intermediate representation (IR) code with
  as little boilerplate as possible
- compilation targeting the host platform, running a comprehensive set
  of optimization passes
- minimal interface to execute the compiled FilterX blocks during
  pipeline execution
- methods to debug and trace IR generation, JIT execution (TODO)

The current implementation is not "just-in-time compilation" in the
classical sense, as all FilterX blocks are compiled and optimized
during config initialization, not at runtime.
This can easily be changed in the future.

For now, FilterXJIT is called "JIT" because it uses LLVM JIT
APIs to produce in-memory code.

Signed-off-by: László Várady <laszlo.varady@anno.io>